backend/apps/auth/pyproject.toml
[tool.poetry]
name = "src"
version = "0.1.0"
description = ""
authors = ["Nikita Smirnov", "Pavel Eroshkin", "Alena Konovalova"]
readme = "README.md"

[tool.poetry.scripts]
create-superuser = "src.cli.main:main"

[tool.poetry.dependencies]
python = "~3.11"
redis = "^4.6.0"
fastapi = "^0.99.0"
pydantic = { version = "~1.10", extras = ["dotenv"] }
orjson = "^3.9.1"
uvloop = "^0.17.0"
uvicorn = "^0.22.0"
gunicorn = "^20.1.0"
fastapi-cache2 = { extras = ["redis"], version = "^0.2.1" }
sqlalchemy = { extras = ["asyncio"], version = "~2.0" }
alembic = "^1.11.1"
psycopg = "~3.1"
psycopg-binary = "~3.1"
typer = { extras = ["all"], version = "^0.9.0" }
passlib = { extras = ["bcrypt"], version = "^1.7.4" }
python-jose = { extras = ["cryptography"], version = "^3.3.0" }
python-multipart = "^0.0.6"
fastapi-pagination = "^0.12.6"
opentelemetry-api = "^1.20.0"
opentelemetry-sdk = "^1.20.0"
opentelemetry-instrumentation-fastapi = "^0.41b0"
opentelemetry-exporter-jaeger = "^1.20.0"
authlib = "^1.2.1"
itsdangerous = "^2.1.2"
fastapi-limiter = "^0.1.5"

[tool.poetry.group.dev.dependencies]
pytest = "^7.3.2"
ipython = "^8.13.2"
pre-commit = "^3.3.2"
flake8 = "^6.0.0"
isort = "^5.12.0"
black = "^23.3.0"
types-redis = "^4.5.5.2"
pyproject-flake8 = "^6.0.0.post1"
dlint = "^0.14.1"
flake8-2020 = "^1.8.0"
flake8-bugbear = "^23.5.9"
flake8-builtins = "^2.1.0"
flake8-comprehensions = "^3.12.0"
flake8-deprecated = "^2.0.1"
flake8-variables-names = "^0.0.5"
mccabe = "^0.7.0"
pep8-naming = "^0.13.3"
flake8-pyproject = "^1.2.3"
mypy = "^1.3.0"
pytest-asyncio = "^0.21.0"
httpx = "^0.24.1"
ipykernel = "^6.24.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
